requirements.txt
fastapi==0.100.0
uvicorn[standard]==0.23.0
pydantic==2.12.5
python-dotenv==1.0.0
google-genai==1.47.0
